refactor: organize types into subdirectory for better modularity
Implements Refactor 04: Organize Types into Subdirectory

Changes:
- Create src/types/ directory with focused type modules
- Split monolithic types.ts into specialized files:
  - core.ts: PromptContext, DefinitionProxy, LastToolInfo, Resettable
  - collections.ts: ToolCollection, SystemCollection, VariableCollection
  - tools.ts: ToolOptions, ToolEventCallback, ToolCallbackResult
  - agents.ts: AgentOptions
  - effects.ts: Effect, StepModifier, StepModifications
  - plugins.ts: Plugin, PluginMethod, MergePlugins, PromptWithPlugins
  - index.ts: Re-exports all types for backward compatibility
- Update src/index.ts to import from types/index
- Delete old monolithic src/types.ts file

Benefits:
- Better organization by domain/concern
- Easier navigation and maintenance
- Smaller, focused files are easier to understand
- Import paths remain backward compatible
- All tests pass (227 tests)
